# QuickStart/Fungi_api_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fungi API",
    "description": "API for Fungi project",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/Authorization/LoginUser": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Login user",
        "description": "Performs the user login procedure",
        "operationId": "LoginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Authorization/RegisterUser": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Register user",
        "description": "Performs user registration procedure",
        "operationId": "RegisterUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceptionDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "errorGroup": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LoginUserRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 32,
            "minLength": 8,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginUserResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoginUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "RegisterUserRequest": {
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "type": "string"
          },
          "email": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 32,
            "minLength": 8,
            "pattern": "^(?=.*\\d)(?=.*[A-Z])(?=.*[a-z])(?=.*[!@#$%^&*(),.?\"{}|<>]).+$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegisterUserResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegisterUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegisterUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}